pyproject.toml
[tool.poetry]
name = "payroll"
version = "0.1.0"
description = ""
authors = ["khongtrunght <khongtrunght@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
sqlalchemy = "^2.0.30"
python-jose = "^3.3.0"
pydantic = "^2.7.2"
bcrypt = "^4.1.3"
pydantic-settings = "^2.2.1"
alembic = "^1.13.1"
psycopg = "^3.1.19"
sqlalchemy-utils = "^0.41.2"
click = "^8.1.7"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.1"
pytest = "^8.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
